Moved to utils/once
