+++ utils/once/once_with_error.go
@@ -12,31 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package core_test
+package oncewitherror
 
-import (
-	"errors"
-	"testing"
+import "sync"
 
-	"github.com/ligato/cn-infra/core"
-	"github.com/onsi/gomega"
-)
-
-const (
-	testErrorString = "This is a test error"
-)
-
-func returnErr() error {
-	err := errors.New(testErrorString)
-	return err
+// OnceWithError is a wrapper around sync.Once that properly handles:
+// func() error
+// instead of just
+// func()
+type OnceWithError struct {
+	once sync.Once
+	err  error
 }
 
-func TestBasicUsage(t *testing.T) {
-	gomega.RegisterTestingT(t)
-	owe := &core.OnceWithError{}
-	err := owe.Do(func() error {
-		return returnErr()
+// Do provides the same functionality as sync.Once.Do(func()) but for
+// func() error
+func (owe *OnceWithError) Do(f func() error) error {
+	owe.once.Do(func() {
+		owe.err = f()
 	})
-	gomega.Expect(err).ShouldNot(gomega.BeNil())
-	gomega.Expect(err.Error()).Should(gomega.Equal(testErrorString))
+	return owe.err
 }
